requirements.txt
fastapi
uvicorn
python-dotenv
crewai
mcp
google-genai
pydantic
aiohttp
